refactor: api 함수 형식 통일
